split files into smaller chunks
