+++ __tests__/componentLoader.spec.tsx
@@ -2,7 +2,7 @@ import '@theuiteam/lib-builder/configs/setupEnzyme';
 import { mount } from 'enzyme';
 import * as React from 'react';
 
-import { toLoadable } from '../src/loadable/loadable';
+import { toLoadable } from '../src/loadable/toLoadable';
 import { ImportedComponent } from '../src/ui/Component';
 import loader from '../src/ui/HOC';
