+++ src/entrypoints/boot.ts
@@ -1,7 +1,8 @@
 import { setConfiguration } from '../configuration/config';
-import { assignImportedComponents, done as whenComponentsReady } from '../loadable/loadable';
+import { assignImportedComponents } from '../loadable/assignImportedComponents';
 import { loadByChunkname } from '../loadable/loadByChunkName';
 import { rehydrateMarks } from '../loadable/marks';
+import { done as whenComponentsReady } from '../loadable/pending';
 import { addPreloader } from '../loadable/preloaders';
 import { injectLoadableTracker } from '../trackers/globalTracker';
