+++ src/loadable/loadable.ts
@@ -1,187 +1,9 @@
-import { settings } from '../configuration/config';
-import { AnyFunction, DefaultImport, Loadable, Mark, MarkMeta, Promised } from '../types';
-import { isBackend } from '../utils/detectBackend';
+import { DefaultImport, Loadable } from '../types';
 import { getFunctionSignature, importMatch } from '../utils/signatures';
-import { assingLoadableMark } from './marks';
-import { getPreloaders } from './preloaders';
-
-export interface InnerLoadable<T> extends Loadable<T> {
-  ok: boolean;
-  promise: Promise<T> | undefined;
-
-  _probeChanges(): Promise<boolean>;
-}
-
-let pending: Array<Promise<any>> = [];
-
-const LOADABLE_WEAK_SIGNATURE = new WeakMap<any, Loadable<any>>();
-const LOADABLE_SIGNATURE = new Map<string, Loadable<any>>();
-
-const addPending = (promise: Promise<any>) => pending.push(promise);
-const removeFromPending = (promise: Promise<any>) => (pending = pending.filter(a => a !== promise));
-
-const toKnownSignature = (signature: string, marks: string[]) =>
-  (!settings.checkSignatures && marks.join('|')) || signature;
-
-export function toLoadable<T>(firstImportFunction: Promised<T>, autoImport = true): Loadable<T> {
-  let importFunction = firstImportFunction;
-  const loadImportedComponent = (): Promise<T> =>
-    Promise.all([importFunction(), ...getPreloaders()]).then(([result]) => result);
-  const functionSignature = getFunctionSignature(importFunction);
-  const mark = importMatch(functionSignature);
-
-  let resolveResolution: AnyFunction;
-  const resolution = new Promise<void>(r => {
-    resolveResolution = r;
-  });
-
-  const loadable: InnerLoadable<T> = {
-    // importFunction,
-    mark,
-    resolution,
-    done: false,
-    ok: false,
-    error: null,
-    payload: undefined,
-    promise: undefined,
-
-    isLoading() {
-      return !!this.promise && !this.done;
-    },
-
-    reset() {
-      this.done = false;
-      this.ok = true;
-      this.payload = undefined;
-      this.promise = undefined;
-    },
-
-    replaceImportFunction(newImportFunction) {
-      importFunction = newImportFunction;
-    },
-
-    get importer() {
-      return importFunction;
-    },
-
-    then(cb, err) {
-      if (this.promise) {
-        return this.promise.then(cb, err);
-      }
-      if (err) {
-        err();
-      }
-      return Promise.reject();
-    },
-
-    loadIfNeeded() {
-      if (this.error) {
-        this.reset();
-      }
-      if (!this.promise) {
-        this.load();
-      }
-      return this.promise!;
-    },
-
-    tryResolveSync(then) {
-      if (this.done) {
-        const result = then(this.payload as any);
-        return {
-          then(cb: any) {
-            // synchronous thenable - https://github.com/facebook/react/pull/14626
-            cb(result);
-            return Promise.resolve(result);
-          },
-        } as any;
-      }
-
-      return this.loadIfNeeded().then(then);
-    },
-
-    reload() {
-      if (this.promise) {
-        this.promise = undefined;
-
-        return this.load() as any;
-      }
-      return Promise.resolve();
-    },
-
-    _probeChanges() {
-      return Promise.resolve(importFunction())
-        .then(payload => payload !== this.payload)
-        .catch(err => {
-          throw err;
-        });
-    },
-
-    load() {
-      if (!this.promise) {
-        const promise = (this.promise = loadImportedComponent().then(
-          payload => {
-            this.done = true;
-            this.ok = true;
-            this.payload = payload;
-            this.error = null;
-            removeFromPending(promise);
-            resolveResolution(payload);
-            return payload;
-          },
-          err => {
-            this.done = true;
-            this.ok = false;
-            this.error = err;
-            removeFromPending(promise);
-            throw err;
-          }
-        ));
-        addPending(promise);
-      }
-      return this.promise;
-    },
-  };
-
-  if (mark && mark.length) {
-    LOADABLE_SIGNATURE.set(toKnownSignature(functionSignature, mark), loadable);
-    assingLoadableMark(mark, loadable);
-  } else {
-    if (process.env.NODE_ENV !== 'development') {
-      // tslint:disable-next-line:no-console
-      console.warn(
-        'react-imported-component: no mark found at',
-        importFunction,
-        'Please check babel plugin or macro setup, as well as imported-component\'s limitations. See https://github.com/theKashey/react-imported-component/issues/147'
-      );
-    }
-  }
-
-  // trigger preload on the server side
-  if (isBackend && autoImport) {
-    loadable.load();
-  }
-
-  return loadable;
-}
-
-let readyFlag = false;
-export const isItReady = () => readyFlag;
-
-/**
- * waits for all necessary imports to be fulfilled
- */
-export const done = (): Promise<void> => {
-  if (pending.length) {
-    readyFlag = false;
-    return Promise.all(pending)
-      .then(a => a[1])
-      .then(done);
-  }
-
-  readyFlag = true;
-
-  return Promise.resolve();
-};
+import { done } from './pending';
+import { LOADABLE_SIGNATURE, LOADABLE_WEAK_SIGNATURE } from './registry';
+import { toLoadable } from './toLoadable';
+import { toKnownSignature } from './utils';
 
 /**
  * try to perform a render and loads all chunks required for it
@@ -191,35 +13,6 @@ export const dryRender = (renderFunction: () => void) => {
   renderFunction();
 
   return Promise.resolve().then(done);
-};
-
-export const markMeta: MarkMeta[] = [];
-
-const assignMetaData = (mark: Mark, loadable: Loadable<any>, chunkName: string, fileName: string) => {
-  markMeta.push({ mark, loadable, chunkName, fileName });
-};
-
-type ImportedDefinition = [Promised<any>, string, string, boolean];
-
-/**
- * to be used __only via CLI tools__
- */
-export const assignImportedComponents = (set: ImportedDefinition[]) => {
-  const countBefore = LOADABLE_SIGNATURE.size;
-  set.forEach(imported => {
-    const allowAutoLoad = !(imported[3] || !settings.fileFilter(imported[2]));
-    const loadable = toLoadable(imported[0], allowAutoLoad);
-    assignMetaData(loadable.mark, loadable, imported[1], imported[2]);
-  });
-
-  if (countBefore === LOADABLE_SIGNATURE.size) {
-    // tslint:disable-next-line:no-console
-    console.error('react-imported-component: no import-marks found, please check babel plugin');
-  }
-
-  done();
-
-  return set;
 };
 
 export function executeLoadable(importFunction: DefaultImport<any> | Loadable<any>) {
