+++ src/loadable/utils.ts
@@ -1,0 +1,4 @@
+import { settings } from '../configuration/config';
+
+export const toKnownSignature = (signature: string, marks: string[]) =>
+  (!settings.checkSignatures && marks.join('|')) || signature;
